requirements.txt
flask>=3.0.0
werkzeug>=3.0.1
yt-dlp>=2023.12.30
openai-whisper
torch
torchaudio
ffmpeg-python>=0.2.0
python-docx>=1.0.0
